// projects/customer360/src/app/home/customer/known-agent/models/known-agent.models.ts
import { ICountryCode } from 'src/app/shared/modules/localization/models';

export interface AgentForm {
  items?: Item[];
  collectBankStatements?: boolean;
  collectDeliverOtherBankMail?: boolean;
  collectChequeBooks?: boolean;
  collectCashFromCompanyCheques?: boolean;
  encashmentOfChequesToDefinedLimit?: boolean;
  balanceEnquiry?: boolean;
  submitRequestServiceBranch?: boolean;
  effectiveDate?:Date;
  highRiskIndividual?: string;
  highRiskType?: number;
  idSerialNumber?: string;
  firstName?: string;
  lastName?: string;
  middleName?: string;
  kraPin?: string;
  limit?: number | null;
}

export interface Item {
  code?: ICountryCode;
  mobileNumber?: string;
  email?: string;
  code1?: string;
  code2?: string;
  code3?: string;
  code4?: string;
  code5?: string;
  code6?: string;
}

export interface Country {
  id?: string;
  name?: string;
  icon?: string;
}

export interface IKnownAgent {
  avatarName?: string;
  acctNum: string;
  acctName: string;
  acctType: string;
  custId: string;
  agentName: string;
  email: string;
  phone: string;
  email2?: string;
  phone2?: string;
  kra: string;
  id: string;
  functions: [
    {
      name: string;
      limit: string;
    },
  ];
  effectiveDate: Date;
  endDate: string;
  functionsText: any;
  isBshFunction: boolean;
  dateCreated: string;
  expiryDate?: string;  
  dateRemoved?: string; 
  deleted: string;
  agentLimit?: string;
  accountType?: string;
  ticketId?: string;
  getUploads?: any[];
  firstName?: string;
  lastName?: string;
  backUrl?: string;
  documentIdType?: string;
  currencyCode?: string;
}

// Generated by https://quicktype.io

export interface IKnownAgentFunctions {
  code: string;
  name: string;
  isLimit: string;
  limit: string;
  selected?: boolean;
}